Expose full sync in index.js
